# Add flag to select tfrt backend for CPU.